# Gallery.Api.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gallery API",
    "version": "v1"
  },
  "paths": {
    "/api/articles": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "Gets Articles",
        "description": "Returns a list of Articles.",
        "operationId": "getArticles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Article"
        ],
        "summary": "Creates a new Article",
        "description": "Creates a new Article with the attributes specified\r\n<para />\r\nAccessible only to a ContentDeveloper or an Administrator",
        "operationId": "createArticle",
        "requestBody": {
          "description": "The data used to create the Article",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{cardId}/articles": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "Gets Articles for a Card",
        "description": "Returns a list of Articles based on the Card ID.",
        "operationId": "getCardArticles",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "The id of the Card",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{collectionId}/articles": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "Gets Articles for a Collection",
        "description": "Returns a list of Articles based on the collection ID",
        "operationId": "getCollectionArticles",
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "description": "The id of the Collection",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/exhibits/{exhibitId}/articles": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "Gets Articles for an Exhibit",
        "description": "Returns a list of Articles based on the exhibit's current move and current inject.",
        "operationId": "getExhibitArticles",
        "parameters": [
          {
            "name": "exhibitId",
            "in": "path",
            "description": "The id of the Exhibit",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": [
          "Article"
        ],
        "summary": "Gets a specific Article by id",
        "description": "Returns the Article with the id specified",
        "operationId": "getArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Article",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Article"
        ],
        "summary": "Updates an Article",
        "description": "Updates a Article with the attributes specified.\r\nThe ID from the route MUST MATCH the ID contained in the article parameter\r\n<para />\r\nAccessible only to a ContentDeveloper or an Administrator",
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Id of the Article to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated Article values",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Article"
        ],
        "summary": "Deletes an Article",
        "description": "Deletes an Article with the specified id\r\n<para />\r\nAccessible only to a ContentDeveloper or an Administrator",
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Article to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "dateModified": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "modifiedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "cardId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "move": {
            "type": "integer",
            "format": "int32"
          },
          "inject": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "sourceType": {
            "$ref": "#/components/schemas/SourceType"
          },
          "sourceName": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "datePosted": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ItemStatus": {
        "enum": [
          "Unused",
          "Closed",
          "Critical",
          "Affected",
          "Open"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SourceType": {
        "enum": [
          "News",
          "Social",
          "Email",
          "Phone",
          "Intel",
          "Reporting"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:5000/connect/authorize",
            "tokenUrl": "http://localhost:5000/connect/token",
            "scopes": {
              "openid profile email gallery": "public api access"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "openid profile email gallery"
      ]
    }
  ]
}